requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.24
botocore==1.38.24
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
dataclasses-json==0.6.7
emoji==2.14.1
entrypoints==0.4
Faker==37.3.0
favicon==0.7.0
filelock==3.18.0
filetype==1.2.0
fonttools==4.58.1
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.40.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
htbuilder==0.9.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.62
langchain-google-genai==2.1.5
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.3.42
lxml==5.4.0
Markdown==3.8
markdownlit==0.0.7
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.41.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.6
olefile==0.47
orjson==3.10.18
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
prometheus_client==0.22.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
PyJWT==2.10.1
pymdown-extensions==10.15
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypdf==5.5.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.13.0
selenium==4.33.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==3.15.0
snowflake-snowpark-python==1.32.0
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
st-annotated-text==4.0.2
st-theme==1.2.3
streamlit==1.45.1
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.7.1
streamlit-faker==0.0.3
streamlit-image-coordinates==0.2.1
streamlit-keyup==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.2
tornado==6.5.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
unstructured==0.17.2
unstructured-client==0.35.0
urllib3==2.4.0
validators==0.35.0
watchdog==6.0.0
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zstandard==0.23.0